+++ lua/plugins/init.lua
@@ -2,3 +2,8 @@
 require("plugins.rocks")
 require("plugins.theme")
 
+-- NvPak Home / Control Center
+pcall(function()
+  require("nvpak.home").setup()
+end)
+